requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
qdrant-client
bs4
requests
python-dotenv
tqdm
